backend/requirements.txt
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.4.0
joblib>=1.3.0
flask>=2.3.0,<4.0.0
flask-cors>=4.0.0
python-dotenv
groq
google-generativeai
twilio>=8.0.0
google-cloud-firestore>=2.11.0
firebase-admin>=6.2.0
reportlab>=4.0.0
razorpay>=1.4.0
